chore(clippy): apply cargo clippy --fix machine-applicable suggestions
Workspace-wide auto-fixes from `cargo clippy --fix --workspace --all-targets`:
needless-update, explicit-auto-deref, collapsible-if, redundant-closure,
useless-conversion, len-zero, default-constructed-unit-structs, single-match, and
other machine-applicable lints across 38 files. Behaviour-preserving; first pass of
the clippy cleanup. Remaining warnings (generated-proto + structural lints) handled
next.

+++ crates/tokeira-controller/src/connect_service.rs
@@ -251,7 +251,7 @@ fn encode_routing_update(
             state: Some(bundle_ownership_entry::State::Owner(
                 BundleOwnerMessage {
                     owner_node_id: owner.node_id.0.to_string(),
-                    epoch: owner.epoch.0 as u64,
+                    epoch: owner.epoch.0,
                     ..Default::default()
                 }
                 .into(),
@@ -307,7 +307,7 @@ fn encode_lease_entry(
             state: Some(bundle_ownership_entry::State::Owner(
                 BundleOwnerMessage {
                     owner_node_id: owner.clone(),
-                    epoch: lease.epoch.0 as u64,
+                    epoch: lease.epoch.0,
                     ..Default::default()
                 }
                 .into(),
